Factor four-byte code into *Info equality
Terms declared in multiple (application and local) dictionaries end up with a different UID string, since they are declared in different places. This is by design. However, application dictionaries often redefine certain terms (`application` being the main example). This is problematic because the new term takes precedence over the old one, meaning it compares unequal to the old one, even when they have the same four-byte AE code and should otherwise be considered the same term.

To deal with this, the new *Info equality rules are:
	1. Are the UID strings the same? If so, the *Infos are equal.
	2. Are the AE codes nil? If so, the *Infos are unequal. This means that when no code is specified, a 1:1 equality mapping of UIDs to terms without code is maintained.
	3. Are the AE codes the same? If so, the *Infos are equal. This provision allows seemingly disparate terms with the same AE code to be treated as equal.